restapi/requirements.txt
alembic==0.9.9
amqp==2.2.2
asn1crypto==0.24.0
attrs==17.4.0
base58==0.2.5
billiard==3.5.0.3
boto==2.48.0
celery==4.1.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
enum34==1.1.6
Flask==1.0.2
Flask-Cors==3.0.3
Flask-JWT-Extended==3.7.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
funcsigs==1.0.2
idna==2.6
ipaddress==1.0.19
itsdangerous==0.24
Jinja2==2.10
kombu==4.1.0
Mako==1.0.7
MarkupSafe==1.0
mock==2.0.0
MySQL-python==1.2.5
pbr==3.1.1
Pillow==5.0.0
pluggy==0.6.0
py==1.5.2
pycparser==2.18
pycrypto==2.6.1
pyfcm==1.4.5
PyJWT==1.4.2
PyPDF2==1.26.0
pytest==3.4.2
python-dateutil==2.6.1
python-editor==1.0.3
python-http-client==3.0.0
pytz==2018.3
redis==2.10.6
reportlab==3.4.0
requests>=2.20.0
requests-toolbelt==0.8.0
sendgrid==5.3.0
six==1.11.0
SQLAlchemy==1.2.4
urllib3==1.22
vine==1.1.4
Werkzeug==0.14.1
WTForms==2.1
uuid
Pyrebase==3.0.27
simplejson==3.15.0
pyyaml==3.12
dropbox==9.0.0
algoliasearch==1.17.0
google-cloud-storage==1.13.0
recombee-api-client==2.2.0
slackclient==1.3.0